refactor(userapi): move capabilities handler out of email_recovery
GetCapabilities was created inside email_recovery.go when every flag it
reported came from userEmailEnabled(). Adding model_catalog broke that
premise: a file named for password recovery now owned a service-wide
manifest and the package's only buildinfo import.

The split had already leaked into the tests, with the email flags covered
in email_recovery_test.go and model_catalog in models_test.go, so neither
file named after the endpoint held its full contract.

Move the handler to capabilities.go and fold both tests into
capabilities_test.go. The merged test asserts model_catalog across all
three mail configurations rather than one, which pins down that the
catalog stays advertised on a Home that cannot send mail, and adds
coverage for the server_info block that had none.

No behaviour change: the route registration in handler.go is untouched.
